updates and reorg
