+++ common/src/server/mod.rs
@@ -1,0 +1,5 @@
+pub mod config;
+pub mod enrollment;
+pub mod heartbeat;
+pub mod jobs;
+pub mod webui;
