+++ server/tests/test_data/3482136c-3176-4272-9bd7-b79f025307d6/enroll.json
@@ -5,6 +5,7 @@
     "os_version": "13.2",
     "uptime": 4550,
     "kernel_version": "22.3.0",
+    "ip": "127.0.0.1",
     "cpu": [
         {
             "frequency": 0,
