+++ webui/src/components/jobs/processes.rs
@@ -1,5 +1,5 @@
 use crate::web::server::request_server;
-use common::server::ProcessJob;
+use common::server::jobs::ProcessJob;
 use leptos::{component, logging::error, view, IntoView};
 use reqwest::Method;
 
@@ -13,38 +13,35 @@ pub(crate) fn HostProcesses(procs: Option<ProcessJob>) -> impl IntoView {
     let headers = vec!["Path", "Name", "PID", "PPID", "Start Time", ""];
 
     view! {
-        <div class="overflow-x-auto col-span-full">
-            <table class="table table-zebra">
-                // Table Header
-                <thead>
-                    <tr>
-                        {headers
-                            .into_iter()
-                            .map(|entry| view! { <th>{entry}</th> })
-                            .collect::<Vec<_>>()}
-                    </tr>
-                </thead>
-                // Table Rows
-                <tbody>
-                    {proc_data
-                        .data
-                        .into_iter()
-                        .map(|res| {
-                            view! {
-                                <tr>
-                                    <td>{res.full_path}</td>
-                                    <td>{res.name}</td>
-                                    <td>{res.pid}</td>
-                                    <td>{res.ppid}</td>
-                                    <td>{res.start_time}</td>
-                                    <th>Details</th>
-                                </tr>
-                            }
-                        })
-                        .collect::<Vec<_>>()}
-                </tbody>
-            </table>
-        </div>
+      <div class="overflow-x-auto col-span-full">
+        <table class="table table-zebra">
+          // Table Header
+          <thead>
+            <tr>
+              {headers.into_iter().map(|entry| view! { <th>{entry}</th> }).collect::<Vec<_>>()}
+            </tr>
+          </thead>
+          // Table Rows
+          <tbody>
+            {proc_data
+                .data
+                .into_iter()
+                .map(|res| {
+                    view! {
+                      <tr>
+                        <td>{res.full_path}</td>
+                        <td>{res.name}</td>
+                        <td>{res.pid}</td>
+                        <td>{res.ppid}</td>
+                        <td>{res.start_time}</td>
+                        <th>Details</th>
+                      </tr>
+                    }
+                })
+                .collect::<Vec<_>>()}
+          </tbody>
+        </table>
+      </div>
     }
 }
