+++ webui/src/components/stats/resources.rs
@@ -1,5 +1,5 @@
 use crate::web::server::request_server;
-use common::server::ServerInfo;
+use common::server::webui::ServerInfo;
 use leptos::logging::error;
 use leptos::{component, create_resource, view, IntoView, SignalGet, Transition};
 use reqwest::Method;
@@ -9,94 +9,90 @@ use reqwest::Method;
 pub(crate) fn Resources() -> impl IntoView {
     let info = create_resource(|| {}, move |_| async move { get_info().await });
     view! {
-        <div class="stat shadow">
-            <div class="stat-title">Server CPU Usage</div>
-            <div class="stat-value">
-                <Transition fallback=move || {
-                    view! { <p>"Loading..."</p> }
-                }>
-                    {move || {
-                        info.get()
-                            .map(|res| {
-                                (res.cpu_usage.iter().sum::<f32>() as f64
-                                    / res.cpu_usage.len() as f64) as u64
-                            })
-                    }}
+      <div class="stat shadow">
+        <div class="stat-title">Server CPU Usage</div>
+        <div class="stat-value">
+          <Transition fallback=move || {
+              view! { <p>"Loading..."</p> }
+          }>
+            {move || {
+                info.get()
+                    .map(|res| {
+                        (res.cpu_usage.iter().sum::<f32>() as f64 / res.cpu_usage.len() as f64)
+                            as u64
+                    })
+            }}
 
-                </Transition>
-                %
-            </div>
+          </Transition>
+          %
         </div>
-        <div class="stat shadow">
-            <div class="stat-title">Server Memory Usage</div>
-            <div class="stat-value">
-                <Transition fallback=move || {
-                    view! { <p>"Loading..."</p> }
-                }>
-                    {move || info.get().map(|res| { res.memory_used / (1024 * 1024 * 1024) })}
-                </Transition>
-                GB
-            </div>
-            <div classs="stat-desc">
-                <Transition fallback=move || {
-                    view! { <p>"Loading..."</p> }
-                }>
-                    {move || info.get().map(|res| { res.total_memory / (1024 * 1024 * 1024) })}
-                </Transition>
-                GB of Total Memory
-            </div>
+      </div>
+      <div class="stat shadow">
+        <div class="stat-title">Server Memory Usage</div>
+        <div class="stat-value">
+          <Transition fallback=move || {
+              view! { <p>"Loading..."</p> }
+          }>{move || info.get().map(|res| { res.memory_used / (1024 * 1024 * 1024) })}</Transition>
+          GB
         </div>
-        <div class="stat shadow">
-            <div class="stat-title">Server Disk Usage</div>
-            <div class="stat-value">
-                <Transition fallback=move || {
-                    view! { <p>"Loading..."</p> }
-                }>
-                    {move || {
-                        info.get()
-                            .map(|res| {
-                                let mut usage = 0;
-                                for disk in res.disk_info {
-                                    if disk.disk_usage > usage {
-                                        usage = disk.disk_usage;
-                                    }
-                                }
-                                usage / (1000 * 1000 * 1000)
-                            })
-                    }}
+        <div classs="stat-desc">
+          <Transition fallback=move || {
+              view! { <p>"Loading..."</p> }
+          }>{move || info.get().map(|res| { res.total_memory / (1024 * 1024 * 1024) })}</Transition>
+          GB of Total Memory
+        </div>
+      </div>
+      <div class="stat shadow">
+        <div class="stat-title">Server Disk Usage</div>
+        <div class="stat-value">
+          <Transition fallback=move || {
+              view! { <p>"Loading..."</p> }
+          }>
+            {move || {
+                info.get()
+                    .map(|res| {
+                        let mut usage = 0;
+                        for disk in res.disk_info {
+                            if disk.disk_usage > usage {
+                                usage = disk.disk_usage;
+                            }
+                        }
+                        usage / (1000 * 1000 * 1000)
+                    })
+            }}
 
-                </Transition>
-                GB
-            </div>
-            <div classs="stat-desc">
-                <Transition fallback=move || {
-                    view! { <p>"Loading..."</p> }
-                }>
-                    {move || {
-                        info.get()
-                            .map(|res| {
-                                let mut size = 0;
-                                for disk in res.disk_info {
-                                    if disk.disk_size > size {
-                                        size = disk.disk_size;
-                                    }
-                                }
-                                size / (1000 * 1000 * 1000)
-                            })
-                    }}
+          </Transition>
+          GB
+        </div>
+        <div classs="stat-desc">
+          <Transition fallback=move || {
+              view! { <p>"Loading..."</p> }
+          }>
+            {move || {
+                info.get()
+                    .map(|res| {
+                        let mut size = 0;
+                        for disk in res.disk_info {
+                            if disk.disk_size > size {
+                                size = disk.disk_size;
+                            }
+                        }
+                        size / (1000 * 1000 * 1000)
+                    })
+            }}
 
-                </Transition>
-                GB Total Disk Size
-            </div>
+          </Transition>
+          GB Total Disk Size
         </div>
-        <div class="stat shadow">
-            <div class="stat-title">Server Uptime in Seconds</div>
-            <div class="stat-value">
-                <Transition fallback=move || {
-                    view! { <p>"Loading..."</p> }
-                }>{move || info.get().map(|res| { res.uptime })}</Transition>
-            </div>
+      </div>
+      <div class="stat shadow">
+        <div class="stat-title">Server Uptime in Seconds</div>
+        <div class="stat-value">
+          <Transition fallback=move || {
+              view! { <p>"Loading..."</p> }
+          }>{move || info.get().map(|res| { res.uptime })}</Transition>
         </div>
+      </div>
     }
 }
