+++ webui/src/components/stats/scripts.rs
@@ -4,9 +4,9 @@ use leptos::{component, view, IntoView};
 /// Scripts page
 pub(crate) fn Scripts() -> impl IntoView {
     view! {
-        <div class="stat shadow">
-            <div class="stat-title">Server Scripts</div>
-            <div class="stat-value">0</div>
-        </div>
+      <div class="stat shadow">
+        <div class="stat-title">Server Scripts</div>
+        <div class="stat-value">0</div>
+      </div>
     }
 }
