+++ webui/src/ui/endpoints.rs
@@ -10,12 +10,12 @@ use leptos_meta::Stylesheet;
 /// List endpoints page
 pub(crate) fn Endpoints() -> impl IntoView {
     view! {
-        <Stylesheet id="leptos" href="/pkg/tailwind.css"/>
-        <div>
-            <Menu/>
-            <Enrollment/>
-        </div>
-        <Footer/>
+      <Stylesheet id="leptos" href="/pkg/tailwind.css"/>
+      <div>
+        <Menu/>
+        <Enrollment/>
+      </div>
+      <Footer/>
     }
 }
 
@@ -23,10 +23,10 @@ pub(crate) fn Endpoints() -> impl IntoView {
 /// Endpoint info page
 pub(crate) fn EndpointInfo() -> impl IntoView {
     view! {
-        <Stylesheet id="leptos" href="/pkg/tailwind.css"/>
-        <div class="grid grid-cols-3">
-            <Menu/>
-            <GetInfo/>
-        </div>
+      <Stylesheet id="leptos" href="/pkg/tailwind.css"/>
+      <div class="grid grid-cols-3">
+        <Menu/>
+        <GetInfo/>
+      </div>
     }
 }
