+++ webui/src/ui/home.rs
@@ -1,4 +1,4 @@
-use common::server::EndpointOS;
+use common::server::webui::EndpointOS;
 use leptos::{component, view, IntoView};
 use leptos_meta::Stylesheet;
 
@@ -21,22 +21,22 @@ pub(crate) fn Home() -> impl IntoView {
     );
 
     view! {
-        <Stylesheet id="leptos" href="/pkg/tailwind.css"/>
-        <div class="grid grid-cols-4">
-            <Menu/>
-            <div class="col-span-1">
-                <Stats os=EndpointOS::All html=all/>
-            </div>
-            <div class="col-span-1">
-                <Stats os=EndpointOS::Linux html=linux/>
-            </div>
-            <div class="col-span-1">
-                <Stats os=EndpointOS::MacOS html=macos/>
-            </div>
-            <div class="col-span-1">
-                <Stats os=EndpointOS::Windows html=windows/>
-            </div>
+      <Stylesheet id="leptos" href="/pkg/tailwind.css"/>
+      <div class="grid grid-cols-4">
+        <Menu/>
+        <div class="col-span-1">
+          <Stats os=EndpointOS::All html=all/>
         </div>
-        <Footer/>
+        <div class="col-span-1">
+          <Stats os=EndpointOS::Linux html=linux/>
+        </div>
+        <div class="col-span-1">
+          <Stats os=EndpointOS::MacOS html=macos/>
+        </div>
+        <div class="col-span-1">
+          <Stats os=EndpointOS::Windows html=windows/>
+        </div>
+      </div>
+      <Footer/>
     }
 }
